+++ fuzz/fuzz_targets/fuzz_xml.rs
@@ -10,7 +10,7 @@ use xmlparser::FromSpan;
 fuzz_target!(|data: &[u8]| {
     if let Ok(text) = str::from_utf8(data) {
         let mut n = 0;
-        for _ in xmlparser::Tokenizer::from_str(text) {
+        for _ in xmlparser::Tokenizer::from(text) {
             n += 1;
 
             if n == 1000 {