+++ src/lib.rs
@@ -2,7 +2,7 @@
 A low-level [XML 1.0](https://www.w3.org/TR/xml/) parser implementation.
 
 ```rust
-for token in xmlparser::Tokenizer::from_str("<tagname name='value'/>") {
+for token in xmlparser::Tokenizer::from("<tagname name='value'/>") {
     println!("{:?}", token);
 }
 ```